+++ spirit-log/examples/logging.rs
@@ -1,8 +1,9 @@
-extern crate serde;
+#[macro_use]
+extern crate log;
+extern crate spirit;
+extern crate spirit_log;
 #[macro_use]
 extern crate serde_derive;
-extern crate spirit;
-extern crate spirit_daemonize;
 #[macro_use]
 extern crate structopt;
 
@@ -10,17 +11,17 @@ use std::thread;
 use std::time::Duration;
 
 use spirit::Spirit;
-use spirit_daemonize::{Daemon, DaemonOpts};
+use spirit_log::{Cfg as LogCfg, Opts as LogOpts};
 
 #[derive(Clone, Debug, StructOpt)]
 struct Opts {
     #[structopt(flatten)]
-    daemon: DaemonOpts,
+    log: LogOpts,
 }
 
 impl Opts {
-    fn daemon(&self) -> &DaemonOpts {
-        &self.daemon
+    fn log(&self) -> LogOpts {
+        self.log.clone()
     }
 }
 
@@ -32,24 +33,29 @@ struct Ui {
 
 #[derive(Clone, Debug, Default, Deserialize)]
 struct Cfg {
-    #[serde(default)]
-    daemon: Daemon,
+    #[serde(flatten)]
+    log: LogCfg,
     ui: Ui,
 }
 
 impl Cfg {
-    fn daemon(&self) -> Daemon {
-        self.daemon.clone()
+    fn log(&self) -> LogCfg {
+        self.log.clone()
     }
 }
 
 const DEFAULT_CONFIG: &str = r#"
-[daemon]
-pid_file = "/tmp/go_background.pid"
-workdir = "/"
+[[logging]]
+level = "INFO"
+type = "stderr"
+
+[[logging]]
+level = "DEBUG"
+type = "file"
+filename = "/tmp/example.log"
 
 [ui]
-msg = "Hello world"
+msg = "Hello!"
 sleep_ms = 100
 "#;
 
@@ -57,15 +63,11 @@ fn main() {
     Spirit::<Opts, Cfg>::new()
         .config_defaults(DEFAULT_CONFIG)
         .config_exts(&["toml", "ini", "json"])
-        .config_helper(
-            Cfg::daemon,
-            spirit_daemonize::with_opts(Opts::daemon),
-            "daemon",
-        )
+        .config_helper(Cfg::log, Opts::log, "logging")
         .run(|spirit| {
             while !spirit.is_terminated() {
                 let cfg = spirit.config();
-                println!("{}", cfg.ui.msg);
+                info!("{}", cfg.ui.msg);
                 thread::sleep(Duration::from_millis(cfg.ui.sleep_ms));
             }
             Ok(())