+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Daemonization extracted to separate helper crate (`spirit-daemonize`).
+* Logging extracted to a separate helper crate (`spirit-log`).
+* Daemonization extracted to a separate helper crate (`spirit-daemonize`).
 * Ability not to start the background thread.
 * Dropped the `S` type parameter. It now keeps the config internally, but if
   exporting to a global variable is needed, a `helpers::cfg_store` helper is